+++ SDK V3.X/Bot/BotBroadcaster/BotBroadcaster.cs
@@ -1,5 +1,4 @@
 ﻿using BotBroadcaster.Model;
-using BotLibrary;
 using BotLibrary.Model;
 using Rainbow;
 using Rainbow.Medias;
@@ -7,12 +6,8 @@ using Rainbow.Model;
 using Rainbow.WebRTC;
 using Rainbow.WebRTC.Abstractions;
 using Rainbow.WebRTC.Desktop;
-using Stateless.Graph;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;