+++ SDK V3.X/Bot/BotBroadcaster/Model/StreamAndDevice.cs
@@ -1,10 +1,6 @@
 ﻿using Rainbow.Medias;
 using Rainbow.WebRTC.Abstractions;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BotBroadcaster.Model
 {