+++ SDK V3.X/Bot/BotBroadcaster/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 # Rainbow CSharp SDK v3 - Bot Broadcaster
 
-##  [0.0.2] - 2025-02-XX
+##  [0.0.2]
+- Add **bot** object in JSON to specify which bot must use the configuration - see chapter [File botConfiguration.json](../README.md#botConfiguration.json) for details.
+- **ApplicationMessage**, to update the bot configuration, must have the first **XmlElement** with **"botConfiguration"** as name - see chapter [Using ApplicationMessage](../README.md#UsingApplicationMessage) for details.
+
+##  [0.0.1.3]
 Configuration structure accepted - see chapter [Extended BotConfiguration](../README.md#ExtendedBotConfiguration) for details.
 
 - **administrators** object, **bubbleInvitationAutoAccept** and **userInvitationAutoAccept**: